SCHED-298: Add items to panel
