+++ src/components/DatePicker/DatePicker.tsx
@@ -9,13 +9,14 @@ export default function DatePicker(){
    
 
     return(<>
-             <h5>Calendar</h5>
+            
             <Calendar id="range" 
                 value={dates2} 
                 onChange={(e) => setDates2(e.value)} 
                 selectionMode="range" 
                 readOnlyInput 
-                showButtonBar/>
+                showButtonBar
+                showIcon/>
        
             </>
     );
